requirements.txt
openai==1.9.0
gradio==4.22.0
python-dotenv==0.21.0
pretty_errors==1.2.25
sxtwl
ichingshifa
ephem-4.1.5
cn2an